Restructure into packagable format
This commit moves much of the content of the repository into a
subdirectory named `easyucs` that is a Python package. It adds a
`setup.py` in the root of the repository which contains metadata
necessary to build, distribute, and install EasyUCS as a Python package.

Import statements throughout the project are adjusted to reflect the
package structure: `import common` is changed to `import
easyucs.common`, and `from __init__ import ...` style imports are
changed to just `from easyucs import ...`.

(To note: some of the `from __init__ import`s may no longer be necessary
in this format.)

The scripts `easyucs.py` and `easyucs_gui.py` are also made part of the
`easyucs` package, and adjusted to act as modules. They can be invoked
like `python -m easyucs.easyucs <args>` or
`python -m easyucs.easyucs_gui <args>` respectively. Their imports are
changed to respect the package structure, and it is now more robust for
`easyucs_gui.py` to import from `easyucs.py`.